demo-project/requirements.txt
django-json-logging==0.0.3
